+++ packages/core-js/actual/json/is-raw-json.js
@@ -1,5 +1,5 @@
 'use strict';
+var parent = require('../../stable/json/is-raw-json');
 require('../../modules/esnext.json.is-raw-json');
-var path = require('../../internals/path');
 
-module.exports = path.JSON.isRawJSON;
+module.exports = parent;