+++ tests/unit-pure/es.json.raw-json.js
@@ -1,6 +1,6 @@
 import { FREEZING } from '../helpers/constants.js';
-import rawJSON from 'core-js-pure/actual/json/raw-json';
-import stringify from 'core-js-pure/actual/json/stringify';
+import rawJSON from 'core-js-pure/es/json/raw-json';
+import stringify from 'core-js-pure/es/json/stringify';
 import isFrozen from 'core-js-pure/es/object/is-frozen';
 import hasOwn from 'core-js-pure/es/object/has-own';